create pipelines submodule and move pipeline files
