+++ src/aidapta/pipelines/layout_ocr.py
@@ -1,12 +1,17 @@
 """A pipeline for extracting metadata from MODS files and imges from PDF files.
+It applyes image search and analysis based  in two steps:
+    First, it analyses the layout of the PDF file using the pdfminer.six library.
+    Second, it applies OCR to the pages where no images were found by layout analysis.
 Author: Manuel Garcia
 """
 
+import hydra
 import os
 import pathlib
 import shutil
 import time
 import logging
+import aidapta.ocr as ocr
 from pdfminer.high_level import extract_pages
 from pdfminer.image import ImageWriter
 from tqdm import tqdm
@@ -14,51 +19,40 @@ from aidapta.utils import extract_mods_metadata, get_entry_number_from_mods
 from aidapta.captions import find_caption_by_distance, find_caption_by_text
 from aidapta.layout import sort_layout_elements, create_output_dir
 from aidapta.metadata import Document, Metadata, Visual, FilePath
+from omegaconf import DictConfig
+from aidapta.captions import OffsetDistance
 
-
-def main(entry_id: str,):
+@hydra.main(version_base=None, config_path="", config_name="config")
+def main(cfg: DictConfig) -> None:
 
     start_time = time.time()
 
-    ###########################################
-    ## SETTINGS                              ##
-    ###########################################
+    #SETTINGS                              
 
     # SELECT INPUT DIRECTORY
-    INPUT_DIR = "data-pipelines/data/design-data100/" # an absolute path is recommended,
+    INPUT_DIR = cfg.dir.input # an absolute path is recommended,
 
     # SELECT OUTPUT DIRECTORY
     # if run multiple times to the same output directory, the images will be duplicated and 
     # metadata will be overwritten
     # This will become the root directory for a Visual object
-    OUTPUT_DIR = INPUT_DIR
+    OUTPUT_DIR = cfg.dir.output # an absolute path is recommended   
     # SET MODS FILE
+    entry_id = cfg.entry
     MODS_FILE = os.path.join(INPUT_DIR, entry_id+"_mods.xml")
 
     # TEMPORARY DIRECTORY
     # this directory is used to store temporary files. PDF files are copied to this directory
-    TMP_DIR = "data-pipelines/data/tmp/" 
-
-    # CAPTION MATCH SETTINGS
-    CAP_SETTINGS ={"method": "bbox",
-            "offset": 14, # one unit equals 1/72 inch or 0.3528 mm
-            "direction": "down", # all directions
-            "keywords": ['figure', 'caption', 'figuur'] # no case sentitive
-            }
-
+    TMP_DIR = cfg.dir.temp 
+    
     # Create output directory for the entry
     entry_directory = create_output_dir(OUTPUT_DIR, entry_id)
-
+    
     # start logging
     logging.basicConfig(filename=os.path.join(OUTPUT_DIR, entry_id, entry_id + '.log'), encoding='utf-8', level=logging.INFO)
     logging.info("Starting pipeline for entry: " + entry_id)
 
-    # SETTINGS FOR THE IMAGE EXTRACTION
-    IMG_SETTINGS = {"width": 100, "height": 100} # recommended values: 0, 0
-
-    ###########################################
-
-    logging.info("Image settings: " + str(IMG_SETTINGS))
+    logging.info("Image settings: " + str(cfg.layout.image_settings))
 
     # EXTRACT METADATA FROM MODS FILE
     meta_blob = extract_mods_metadata(MODS_FILE)
@@ -77,6 +71,9 @@ def main(entry_id: str,):
     base_url = "http://resolver.tudelft.nl/" 
     entry.add_web_url(base_url)
 
+    offset_dist = OffsetDistance ( cfg.layout.caption_settings.offset[0], cfg.layout.caption_settings.offset[1])
+    print ("offset distance", offset_dist)
+
     # PROCESS PDF FILES
     start_processing_time = time.time()
     for pdf in PDF_FILES:
@@ -88,35 +85,46 @@ def main(entry_id: str,):
 
         # PREPARE OUTPUT DIRECTORY
         pdf_file_name = pathlib.Path(pdf_document.location).stem
-        image_directory = create_output_dir( os.path.join(OUTPUT_DIR, entry_directory), pdf_file_name) 
+        image_directory = create_output_dir(entry_directory, 
+                                            pdf_file_name) # returns a pathlib object
+        
+        ocr_directory = create_output_dir(image_directory, "ocr")
 
         # PROCESS SINGLE PDF 
         pdf_pages = extract_pages(pdf_document.location)
 
         pages = []
         for page in tqdm(pdf_pages, desc="Reading pages", unit="pages"):
-            elements = sort_layout_elements(page, img_height=IMG_SETTINGS["width"], img_width=IMG_SETTINGS["height"])
+            elements = sort_layout_elements(page, img_height = cfg.layout.image_settings.width, 
+                                            img_width=cfg.layout.image_settings.height)
             pages.append(elements)
 
-        for page in tqdm(pages, desc="Extracting images", total=len(pages), unit="pages"):
+        ocr_pages = []
+        # PROCESS PAGE USING LAYOUT ANALYSIS
+        for page in tqdm(pages, desc="layout analysis", total=len(pages), 
+                         unit="sorted pages"):
 
-            iw = ImageWriter( os.path.join(OUTPUT_DIR, entry_directory, image_directory) )
+            iw = ImageWriter(image_directory)
+
+            if page["images"] == []: # collects pages where no images were found by layout analysis
+                ocr_pages.append(page)
         
             for img in page["images"]:
             
-                visual = Visual(document_page=page["page_number"], document=pdf_document, bbox=img.bbox)
+                visual = Visual(document_page=page["page_number"], 
+                                document=pdf_document, bbox=img.bbox)
                 
-                # Search for captions using proximity to image Bboxes
-                # This might generate multiple matches
+                # Search for captions using proximity to image
+                # This may generate multiple matches
+                print("offset type", type(offset_dist))
                 bbox_matches =[]
                 for _text in page["texts"]:
-                    match = find_caption_by_distance(img, _text, offset_distance=CAP_SETTINGS["offset"], 
-                                                direction=CAP_SETTINGS["direction"])
+                    match = find_caption_by_distance(img, _text, offset= offset_dist, 
+                                                direction= cfg.layout.caption_settings.direction)
                     if match:
                         bbox_matches.append(match)
-                # Search for captions using text analysis (keywords)
-                # if more than one bbox matches are found
-                if len(bbox_matches) == 0:
+                # Search for captions using proximity (offset) and text analyses (keywords)
+                if len(bbox_matches) == 0: # if more than one bbox matches, move to text analysis
                     pass # don't set any caption
                 elif len(bbox_matches) == 1:
                     caption = ""
@@ -125,25 +133,26 @@ def main(entry_id: str,):
                     visual.set_caption(caption)
                 else: # more than one matches in bbox_matches
                     for _text in bbox_matches:
-                        text_match = find_caption_by_text(_text, keywords=CAP_SETTINGS["keywords"])
+                        text_match = find_caption_by_text(_text, keywords=cfg.layout.caption_settings.keywords)
                     if text_match:
                         caption = ""
                         for text_line in bbox_matches[0]:
                             caption += text_line.get_text().strip() 
-                    # Set the caption to the firt text match.
+                    # Set the caption to the first text match.
+                    # TODO: implement recording multiple matches when no single match can be ruled out
                     # All other matches will be ignored. 
-                    # This may introduce mistakes
+                    # This may introduce errors, but it is better than having multiple captions
                         try:
                             visual.set_caption(caption)
                         except Warning: # ignore warnings when caption is already set.
                             logging.warning("Caption already set for image: " + img.name)
                             pass
                         
-                # TODO: https://github.com/pdfminer/pdfminer.six/pull/854
                 # rename image name to include page number
                 img.name =  str(entry_id) + "-page" + str(page["page_number"]) + "-" + img.name
                 # save image to file
             
+                # TODO: https://github.com/pdfminer/pdfminer.six/pull/854
                 try:
                     image_file_name =iw.export_image(img) # returns image file name, 
                     # which last part is automatically generated by pdfminer to guarantee uniqueness
@@ -153,14 +162,21 @@ def main(entry_id: str,):
                     pass
                     
                 visual.set_location( FilePath(
-                                        OUTPUT_DIR, os.path.join(
-                                            entry_directory, image_directory, image_file_name
+                                         str(image_directory), image_file_name
                                         ) 
-                                    ) )
+                                    ) 
             
                 # add visual to entry
                 entry.add_visual(visual)
 
+        # PROCESS PAGE USING OCR ANALYSIS
+        for ocr_page in tqdm(ocr_pages, desc="OCR analysis", total=len(ocr_pages), unit="OCR pages"):
+
+            # if page["images"] == []: # apply to pages where no images were found by layout analysis
+                page_image = ocr.convert_pdf_to_image(pdf_document.location, dpi= cfg.ocr.resolution, first_page=ocr_page["page_number"], last_page=ocr_page["page_number"])
+                ocr_results = ocr.extract_bboxes_from_horc(page_image, config='--psm 3 --oem 1', entry_id=entry_id, page_number=ocr_page["page_number"])
+                ocr.mark_bounding_boxes(ocr_results, ocr_directory, filter_size=100)      
+    
     end_processing_time = time.time()
     processing_time = end_processing_time - start_processing_time
     logging.info("PDF processing time: " + str(processing_time))
@@ -174,17 +190,17 @@ def main(entry_id: str,):
     temp_entry_directory = create_output_dir( os.path.join(TMP_DIR, entry_id))
 
     mods_file_name = pathlib.Path(MODS_FILE).stem + ".xml"
-    if not os.path.exists(os.path.join(TMP_DIR, temp_entry_directory, mods_file_name)):
-        shutil.copy2(MODS_FILE, os.path.join(
-            TMP_DIR, temp_entry_directory, entry_id
-            ))
+    if not os.path.exists(os.path.join(temp_entry_directory, mods_file_name)):
+        shutil.copy2(MODS_FILE, temp_entry_directory)
+
     for pdf in PDF_FILES:
-        if not os.path.exists(os.path.join(TMP_DIR, temp_entry_directory, pdf)):
-            shutil.copy2(pdf, os.path.join(TMP_DIR, temp_entry_directory, entry_id))
+        if not os.path.exists(os.path.join(temp_entry_directory, pdf)):
+            shutil.copy2(pdf, temp_entry_directory)
     
     # SAVE METADATA TO files
-    entry.save_to_json(os.path.join(OUTPUT_DIR, entry_directory, entry_id + "-metadata.json"))
-    entry.save_to_csv(os.path.join(OUTPUT_DIR, entry_directory, entry_id + "-metadata.csv"))
+    # json_file = str(os.path.join(entry_directory, entry_id) + "-metadata.json")
+    csv_file = str(os.path.join(entry_directory, entry_id) + "-metadata.csv")
+    json_file = str(os.path.join(entry_directory, entry_id) + "-metadata.json")
 
     end_time = time.time()
     total_time = end_time - start_time
@@ -194,5 +210,5 @@ def main(entry_id: str,):
 if __name__ == "__main__":
     
     # for id in range(11,12):
-    str_id = str(1).zfill(5)
-    main(str_id)
+    str_id = str(2).zfill(5)
+    main()
